+++ test/app/Http/Controllers/Admin_CommentController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Comment;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Gate;
 
 class Admin_CommentController extends Controller
 {
@@ -16,7 +17,9 @@ class Admin_CommentController extends Controller
      */
     public function index($id)
     {
-
+        if (!Gate::allows("view-product")) {
+            abort(403);
+        }
         $comments = Comment::where('product_id', $id)->get();
         return view('admin.comment.list',[
             'comments' =>$comments
@@ -86,6 +89,9 @@ class Admin_CommentController extends Controller
      */
     public function destroy(Request $request,Comment $comment)
     {
+        if (!Gate::allows("delete-product")) {
+            abort(403);
+        }
         try{
             $msg = 'Delete ID:'. $comment->id. ' comment successfully';
             $comment->delete();
@@ -97,7 +103,11 @@ class Admin_CommentController extends Controller
         }
         return redirect()->route("admin.comment.index");
     }
+
     public function delete(Request $request,Comment $comment, $id){
+        if (!Gate::allows("delete-product")) {
+            abort(403);
+        }
         try{
             $msg = 'Delete ID:'. $comment->id. ' comment successfully';
             Comment::where('id', $id)->delete();
